Added a PERC drive monitoring cronjob script
My DELL server has a PERC RAID controller.  After losing all of the data on the drives to a RAID failure, I decided to be more proactive and receive email alerts for a failed drive.  This script simply looks for something after "DRIVE" that is not "Onln" and "DHS".  If the condition is met, it will write the FAIL_FILE to disk and send an email (in the opposite order, actually).  If a failed drive is NOT found and that file is found, the script assumes that the error condition has been cleared.  It will remove the file and send an "ALL CLEAR" email.